refactor(java-parser): merge Parser and ParserV2 into single Parser
The original Parser (parser.go) was a low-level PoC struct that exposed
raw tree-sitter types. ParserV2 (parser_v2.go) was the production
implementation with caching, logging, and parallel support. Since all
production code uses ParserV2 and both share identical extractClasses
logic, merge them into a single Parser.

Changes:
- Delete parser_v2.go, fold all logic into parser.go
- Rename NewParserV2 -> NewParser, ParserV2 -> Parser
- Extract findJavaFiles() helper to remove duplicated filepath.Walk
- extractAllClasses() replaces duplicate per-struct extractClasses methods
- Delete parser_v2_test.go, rewrite parser_test.go with merged tests
  covering ParseFile, interfaces, cache, no-cache, directory, parallel,
  log levels, and concurrent access
- Update all callers (collector.go, integration tests)
