+++ api-collector-java/parser/parser_test.go
@@ -1,87 +1,58 @@
 package parser
 
 import (
-	"fmt"
+	"os"
+	"path/filepath"
 	"testing"
 )
 
-func TestParseSpringMVCController(t *testing.T) {
-	parser, err := NewParser()
+func TestParser_ParseFile(t *testing.T) {
+	p, err := NewParser(ParserOptions{
+		CacheDir: filepath.Join(t.TempDir(), "cache"),
+		LogLevel: LogLevelError,
+	})
 	if err != nil {
 		t.Fatalf("Failed to create parser: %v", err)
 	}
-	defer parser.Close()
+	defer p.Close()
 
-	// Parse the test file
-	tree, source, err := parser.ParseFile("../testdata/UserController.java")
+	result, err := p.ParseFile("../testdata/UserController.java")
 	if err != nil {
 		t.Fatalf("Failed to parse file: %v", err)
 	}
-	defer tree.Close()
-
-	// Extract classes
-	classes, err := parser.ExtractClasses(tree, source)
-	if err != nil {
-		t.Fatalf("Failed to extract classes: %v", err)
+	if result.Error != nil {
+		t.Fatalf("Parse result error: %v", result.Error)
+	}
+	if len(result.Classes) != 1 {
+		t.Fatalf("Expected 1 class, got %d", len(result.Classes))
 	}
 
-	// Verify we found the UserController class
-	if len(classes) != 1 {
-		t.Fatalf("Expected 1 class, got %d", len(classes))
-	}
-
-	class := classes[0]
-
-	// Verify class name
+	class := result.Classes[0]
 	if class.Name != "UserController" {
 		t.Errorf("Expected class name 'UserController', got '%s'", class.Name)
 	}
-
-	// Verify package name
 	if class.Package != "com.example.demo.controller" {
 		t.Errorf("Expected package 'com.example.demo.controller', got '%s'", class.Package)
 	}
-
-	// Verify class annotations
-	if len(class.Annotations) < 2 {
-		t.Errorf("Expected at least 2 class annotations, got %d", len(class.Annotations))
+	if len(class.Annotations) != 2 {
+		t.Errorf("Expected 2 class annotations, got %d", len(class.Annotations))
 	}
-
-	// Check for @RestController
-	hasRestController := false
-	for _, ann := range class.Annotations {
-		if ann.Name == "RestController" {
-			hasRestController = true
-			break
-		}
-	}
-	if !hasRestController {
-		t.Error("Expected @RestController annotation on class")
-	}
-
-	// Check for @RequestMapping
-	hasRequestMapping := false
-	var requestMappingValue string
-	for _, ann := range class.Annotations {
-		if ann.Name == "RequestMapping" {
-			hasRequestMapping = true
-			requestMappingValue = ann.Params["value"]
-			break
-		}
-	}
-	if !hasRequestMapping {
-		t.Error("Expected @RequestMapping annotation on class")
-	}
-	if requestMappingValue != "/api/users" {
-		t.Errorf("Expected @RequestMapping value '/api/users', got '%s'", requestMappingValue)
-	}
-
-	// Verify methods
 	if len(class.Methods) != 5 {
 		t.Errorf("Expected 5 methods, got %d", len(class.Methods))
 	}
 
-	// Check getUser method
+	// Verify @RestController annotation
+	hasRestController := false
+	for _, ann := range class.Annotations {
+		if ann.Name == "RestController" {
+			hasRestController = true
+		}
+	}
+	if !hasRestController {
+		t.Error("Expected @RestController annotation")
+	}
+
+	// Verify getUser method and its @PathVariable parameter
 	var getUserMethod *Method
 	for i := range class.Methods {
 		if class.Methods[i].Name == "getUser" {
@@ -89,113 +60,198 @@ func TestParseSpringMVCController(t *testing.T) {
 			break
 		}
 	}
-
 	if getUserMethod == nil {
 		t.Fatal("Expected to find getUser method")
 	}
-
-	// Verify @GetMapping annotation
-	hasGetMapping := false
-	var getMappingValue string
-	for _, ann := range getUserMethod.Annotations {
-		if ann.Name == "GetMapping" {
-			hasGetMapping = true
-			getMappingValue = ann.Params["value"]
-			break
-		}
-	}
-	if !hasGetMapping {
-		t.Error("Expected @GetMapping annotation on getUser method")
-	}
-	if getMappingValue != "/{id}" {
-		t.Errorf("Expected @GetMapping value '/{id}', got '%s'", getMappingValue)
-	}
-
-	// Verify method parameters
 	if len(getUserMethod.Parameters) != 1 {
-		t.Errorf("Expected 1 parameter, got %d", len(getUserMethod.Parameters))
+		t.Fatalf("Expected 1 parameter, got %d", len(getUserMethod.Parameters))
 	}
-
-	if len(getUserMethod.Parameters) > 0 {
-		param := getUserMethod.Parameters[0]
-		if param.Name != "id" {
-			t.Errorf("Expected parameter name 'id', got '%s'", param.Name)
-		}
-		if param.Type != "Long" {
-			t.Errorf("Expected parameter type 'Long', got '%s'", param.Type)
-		}
-
-		// Check for @PathVariable annotation
-		hasPathVariable := false
-		for _, ann := range param.Annotations {
-			if ann.Name == "PathVariable" {
-				hasPathVariable = true
-				break
-			}
-		}
-		if !hasPathVariable {
-			t.Error("Expected @PathVariable annotation on id parameter")
+	param := getUserMethod.Parameters[0]
+	if param.Name != "id" || param.Type != "Long" {
+		t.Errorf("Expected param 'Long id', got '%s %s'", param.Type, param.Name)
+	}
+	hasPathVariable := false
+	for _, ann := range param.Annotations {
+		if ann.Name == "PathVariable" {
+			hasPathVariable = true
 		}
 	}
+	if !hasPathVariable {
+		t.Error("Expected @PathVariable on id parameter")
+	}
+}
 
-	// Print summary for manual verification
-	fmt.Println("\n=== Parsing Results ===")
-	fmt.Printf("Class: %s\n", class.Name)
-	fmt.Printf("Package: %s\n", class.Package)
-	fmt.Printf("Class Annotations: %d\n", len(class.Annotations))
-	for _, ann := range class.Annotations {
-		fmt.Printf("  - @%s", ann.Name)
-		if len(ann.Params) > 0 {
-			fmt.Printf("(")
-			first := true
-			for k, v := range ann.Params {
-				if !first {
-					fmt.Printf(", ")
-				}
-				fmt.Printf("%s=\"%s\"", k, v)
-				first = false
-			}
-			fmt.Printf(")")
-		}
-		fmt.Println()
+func TestParser_ParseInterface(t *testing.T) {
+	p, err := NewParser(ParserOptions{LogLevel: LogLevelError})
+	if err != nil {
+		t.Fatalf("Failed to create parser: %v", err)
+	}
+	defer p.Close()
+
+	result, err := p.ParseFile("../testdata/UserClient.java")
+	if err != nil {
+		t.Fatalf("Failed to parse file: %v", err)
+	}
+	if result.Error != nil {
+		t.Fatalf("Parse result error: %v", result.Error)
+	}
+	if len(result.Classes) != 1 {
+		t.Fatalf("Expected 1 class/interface, got %d", len(result.Classes))
 	}
 
-	fmt.Printf("\nMethods: %d\n", len(class.Methods))
-	for _, method := range class.Methods {
-		fmt.Printf("\n  Method: %s\n", method.Name)
-		fmt.Printf("  Return Type: %s\n", method.ReturnType)
-		fmt.Printf("  Annotations: %d\n", len(method.Annotations))
-		for _, ann := range method.Annotations {
-			fmt.Printf("    - @%s", ann.Name)
-			if len(ann.Params) > 0 {
-				fmt.Printf("(")
-				first := true
-				for k, v := range ann.Params {
-					if !first {
-						fmt.Printf(", ")
-					}
-					fmt.Printf("%s=\"%s\"", k, v)
-					first = false
-				}
-				fmt.Printf(")")
+	iface := result.Classes[0]
+	if !iface.IsInterface {
+		t.Error("Expected IsInterface=true")
+	}
+	if iface.Name != "UserClient" {
+		t.Errorf("Expected 'UserClient', got '%s'", iface.Name)
+	}
+}
+
+func TestParser_Cache(t *testing.T) {
+	p, err := NewParser(ParserOptions{
+		CacheDir: filepath.Join(t.TempDir(), "cache"),
+		LogLevel: LogLevelError,
+	})
+	if err != nil {
+		t.Fatalf("Failed to create parser: %v", err)
+	}
+	defer p.Close()
+
+	testFile := "../testdata/UserController.java"
+
+	result1, err := p.ParseFile(testFile)
+	if err != nil {
+		t.Fatalf("First parse failed: %v", err)
+	}
+	result2, err := p.ParseFile(testFile)
+	if err != nil {
+		t.Fatalf("Second parse (cache hit) failed: %v", err)
+	}
+
+	if len(result1.Classes) != len(result2.Classes) {
+		t.Error("Cache returned different number of classes")
+	}
+	if result1.Classes[0].Name != result2.Classes[0].Name {
+		t.Error("Cache returned different class name")
+	}
+
+	if _, err := os.Stat(filepath.Join(p.cache.cacheDir)); os.IsNotExist(err) {
+		t.Error("Cache directory was not created")
+	}
+}
+
+func TestParser_NoCacheMode(t *testing.T) {
+	p, err := NewParser(ParserOptions{LogLevel: LogLevelError})
+	if err != nil {
+		t.Fatalf("Failed to create parser: %v", err)
+	}
+	defer p.Close()
+
+	result, err := p.ParseFile("../testdata/UserController.java")
+	if err != nil {
+		t.Fatalf("Failed to parse file: %v", err)
+	}
+	if result.Error != nil {
+		t.Fatalf("Parse result error: %v", result.Error)
+	}
+	if len(result.Classes) != 1 {
+		t.Fatalf("Expected 1 class, got %d", len(result.Classes))
+	}
+}
+
+func TestParser_ParseDirectory(t *testing.T) {
+	p, err := NewParser(ParserOptions{
+		CacheDir: filepath.Join(t.TempDir(), "cache"),
+		LogLevel: LogLevelError,
+	})
+	if err != nil {
+		t.Fatalf("Failed to create parser: %v", err)
+	}
+	defer p.Close()
+
+	results, err := p.ParseDirectory("../testdata")
+	if err != nil {
+		t.Fatalf("Failed to parse directory: %v", err)
+	}
+	if len(results) == 0 {
+		t.Error("Expected at least one parse result")
+	}
+
+	found := false
+	for _, result := range results {
+		for _, class := range result.Classes {
+			if class.Name == "UserController" {
+				found = true
 			}
-			fmt.Println()
-		}
-		fmt.Printf("  Parameters: %d\n", len(method.Parameters))
-		for _, param := range method.Parameters {
-			fmt.Printf("    - %s %s", param.Type, param.Name)
-			if len(param.Annotations) > 0 {
-				fmt.Printf(" [")
-				for i, ann := range param.Annotations {
-					if i > 0 {
-						fmt.Printf(", ")
-					}
-					fmt.Printf("@%s", ann.Name)
-				}
-				fmt.Printf("]")
-			}
-			fmt.Println()
 		}
 	}
-	fmt.Println("======================")
+	if !found {
+		t.Error("UserController not found in parse results")
+	}
+}
+
+func TestParser_ParseDirectoryParallel(t *testing.T) {
+	p, err := NewParser(ParserOptions{
+		CacheDir: filepath.Join(t.TempDir(), "cache"),
+		LogLevel: LogLevelError,
+	})
+	if err != nil {
+		t.Fatalf("Failed to create parser: %v", err)
+	}
+	defer p.Close()
+
+	parallel, err := p.ParseDirectoryParallel("../testdata", 2)
+	if err != nil {
+		t.Fatalf("Failed to parse directory in parallel: %v", err)
+	}
+	sequential, err := p.ParseDirectory("../testdata")
+	if err != nil {
+		t.Fatalf("Failed to parse directory sequentially: %v", err)
+	}
+	if len(parallel) != len(sequential) {
+		t.Errorf("Parallel returned %d results, sequential returned %d",
+			len(parallel), len(sequential))
+	}
+}
+
+func TestParser_LogLevels(t *testing.T) {
+	for _, level := range []LogLevel{LogLevelDebug, LogLevelInfo, LogLevelWarn, LogLevelError} {
+		p, err := NewParser(ParserOptions{
+			CacheDir: filepath.Join(t.TempDir(), "cache"),
+			LogLevel: level,
+		})
+		if err != nil {
+			t.Fatalf("Failed to create parser with log level %d: %v", level, err)
+		}
+		if _, err := p.ParseFile("../testdata/UserController.java"); err != nil {
+			t.Errorf("Parse failed with log level %d: %v", level, err)
+		}
+		p.Close()
+	}
+}
+
+func TestParser_ConcurrentAccess(t *testing.T) {
+	p, err := NewParser(ParserOptions{
+		CacheDir: filepath.Join(t.TempDir(), "cache"),
+		LogLevel: LogLevelError,
+	})
+	if err != nil {
+		t.Fatalf("Failed to create parser: %v", err)
+	}
+	defer p.Close()
+
+	const n = 50
+	errChan := make(chan error, n)
+	for i := 0; i < n; i++ {
+		go func() {
+			_, err := p.ParseFile("../testdata/UserController.java")
+			errChan <- err
+		}()
+	}
+	for i := 0; i < n; i++ {
+		if err := <-errChan; err != nil {
+			t.Errorf("Concurrent parse failed: %v", err)
+		}
+	}
 }
